Pushing to GitHub before making changes to DIT Group in case of errors
